% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-books.R
\docType{data}
\name{ads}
\alias{ads}
\title{The Art of Data Science}
\format{A \link[tibble:lst]{tibble} representing the book in
\href{http://tidytextmining.com/tidytext.html}{tidy text format} with single word
as token. \link[tidytext:stop_words]{Stop words} are removed.

Data has the following columns:
\itemize{
\item \strong{id} <int> : Index of word inside the book.
\item \strong{book} <chr> : Name of the book.
\item \strong{page} <int> : Page number.
\item \strong{line} <int> : Line number on page (empty lines are ignored).
\item \strong{word} <chr> : Word.
}}
\usage{
ads
}
\description{
Tidied version of the book "The Art of Data Science" by Roger D. Peng and
Elizabeth Matsui published in 2015 - 2017. This book is for sale at
\href{http://leanpub.com/artofdatascience}{leanpub}. It also has its own
\href{https://bookdown.org/rdpeng/artofdatascience/}{web page}.
}
\keyword{datasets}
