% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-books.R
\docType{data}
\name{edas}
\alias{edas}
\title{The Elements of Data Analytic Style}
\format{A \link[tibble:lst]{tibble} representing the book in
\href{http://tidytextmining.com/tidytext.html}{tidy text format} with single word
as token. \link[tidytext:stop_words]{Stop words} are removed.

Data has the following columns:
\itemize{
\item \strong{id} <int> : Index of word inside the book.
\item \strong{book} <chr> : Name of the book.
\item \strong{page} <int> : Page number.
\item \strong{line} <int> : Line number on page (empty lines are ignored).
\item \strong{word} <chr> : Word.
}}
\usage{
edas
}
\description{
Tidied version of the book "The Elements of Data Analytic Style" by Jeff Leek
published in 2014-2015. This book is for sale at
\href{http://leanpub.com/datastyle}{leanpub}.
}
\keyword{datasets}
