% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-files.R
\name{extract_alpha}
\alias{extract_alpha}
\title{Extract alphabetical characters}
\usage{
extract_alpha(x)
}
\arguments{
\item{x}{Vector of strings.}
}
\description{
Extract all alphabetical characters from string.
}
\examples{
# "\\n" is treated as one symbol
extract_alpha(c("\\nword", "\\\\word", " word", "1word2"))

}
