% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-files.R
\name{file_base_name}
\alias{file_base_name}
\title{Get file base name}
\usage{
file_base_name(file)
}
\arguments{
\item{file}{Vector of paths to files.}
}
\description{
Extract file's base name without extension (last three characters preceded by
'.').
}
\examples{
file_base_name(c("some/folder/path/end-file.rmd",
                 "some/folder/path/end-file.Rmd",
                 "some/folder/path/end-file.txt",
                 ".end-file", "end-file."))

}
