% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-books.R
\docType{data}
\name{r4ds}
\alias{r4ds}
\title{R for Data Science}
\format{A \link[tibble:lst]{tibble} representing the book in
\href{http://tidytextmining.com/tidytext.html}{tidy text format} with single word
as token. \link[tidytext:stop_words]{Stop words} are removed.

Data has the following columns:
\itemize{
\item \strong{id} <int> : Index of word inside the book.
\item \strong{book} <chr> : Name of the book.
\item \strong{page} <int> : Page (which is actually a chapter) number.
\item \strong{pageName} <int> : Name of page/chapter.
\item \strong{word} <chr> : Word.
}}
\usage{
r4ds
}
\description{
Tidied version of the book "R for Data Science" by Garrett Grolemund and
Hadley Wickham. This book has its own \href{http://r4ds.had.co.nz/}{web page}.
It is licensed under
\href{http://creativecommons.org/licenses/by-nc-nd/3.0/us/}{Creative Commons Attribution-NonCommercial-NoDerivs 3.0}
United States License.
}
\keyword{datasets}
