% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-files.R
\name{remove_md_emphasis}
\alias{remove_md_emphasis}
\title{Remove markdown emphasis}
\usage{
remove_md_emphasis(x)
}
\arguments{
\item{x}{Character vector.}
}
\description{
Function to remove emphasis from scraped markdown.
}
\examples{
remove_md_emphasis(c("_a", "__a", "*a", "_*a",
                     "a_", "a__", "a*", "a*_",
                     "_a_", "__a__", "*a*", "**a**"))

}
