% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-files.R
\name{tidy_pdf}
\alias{tidy_pdf}
\title{Tidy pdf file}
\usage{
tidy_pdf(file)
}
\arguments{
\item{file}{Path to pdf file.}
}
\value{
A \link[tibble:lst]{tibble} with the following columns:
\itemize{
\item \strong{page} <int> : Word's page number.
\item \strong{line} <int> : Words's line number on page (empty lines are ignored).
\item \strong{word} <chr> : Word.
}
}
\description{
Parse words from \strong{pdf} file into
\href{http://tidytextmining.com/tidytext.html}{tidy format}. It uses
\link[pdftools:pdf_text]{pdftools::pdf_text()} to parse pdf file and
\link[tidytext:unnest_tokens]{tidytext::unnest_tokens()} to produce words from
parsed text.
}
