% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-files.R
\name{tidy_rmd}
\alias{tidy_rmd}
\title{Tidy rmd file}
\usage{
tidy_rmd(file, name = file_base_name(file))
}
\arguments{
\item{file}{Path to pdf file.}

\item{name}{Name of the document represented by file.}
}
\value{
A \link[tibble:lst]{tibble} with the following columns:
\itemize{
\item \strong{name} <int> : Document name.
\item \strong{word} <chr> : Word.
}
}
\description{
Parse words from \strong{rmd} file into
\href{http://tidytextmining.com/tidytext.html}{tidy format}. It reads raw file,
removes some chunks based on regular expressions and uses
\link[tidytext:unnest_tokens]{tidytext::unnest_tokens()} to produce words.
\strong{Note} that regular expressions are design to handle common rmd formatting.
On complicated cases it can give undesired output.
}
